# openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "STT Server API",
        "version": "1.0.0",
        "description": "Speech-to-Text Server mit Transkription, Sprechererkennung und Zusammenfassung"
    },
    "paths": {
...
        "/v1/diarize": {
            "post": {
                "operationId": "v1_diarize_create",
                "summary": "Transcribe with speaker diarization",
                "tags": [
                    "v1"
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/AudioUpload"
                            }
                        },
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/AudioUpload"
                            }
                        }
                    },
                    "required": true
                },
                "security": [
                    {
                        "oauth2": [
                            "write"
                        ]
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DiarizeResponse"
                                }
                            }
                        },
                        "description": ""
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": ""
                    },
                    "413": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": ""
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": ""
                    },
                    "503": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
...
        "/v1/transcribe": {
            "post": {
                "operationId": "v1_transcribe_create",
                "summary": "Transcribe audio file",
                "tags": [
                    "v1"
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/AudioUpload"
                            }
                        },
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/AudioUpload"
                            }
                        }
                    },
                    "required": true
                },
                "security": [
                    {
                        "oauth2": [
                            "write"
                        ]
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TranscribeResponse"
                                }
                            }
                        },
                        "description": ""
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": ""
                    },
                    "413": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": ""
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "AudioUpload": {
                "type": "object",
                "properties": {
                    "file": {
                        "type": "string",
                        "format": "uri",
                        "description": "Audio file (.wav, .mp3, .flac, .ogg, .m4a, .wma, .webm)"
                    },
                    "model": {
                        "type": "string",
                        "default": "small",
                        "description": "Whisper model name"
                    }
                },
                "required": [
                    "file"
                ]
            },
...
            "DiarizeResponse": {
                "type": "object",
                "properties": {
                    "text": {
                        "type": "string"
                    },
                    "diarized_text": {
                        "type": "string"
                    },
                    "segments": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/DiarizedSegment"
                        }
                    }
                },
                "required": [
                    "diarized_text",
                    "segments",
                    "text"
                ]
            },
            "DiarizedSegment": {
                "type": "object",
                "properties": {
                    "speaker": {
                        "type": "string"
                    },
                    "start": {
                        "type": "number",
                        "format": "double"
                    },
                    "end": {
                        "type": "number",
                        "format": "double"
                    },
                    "text": {
                        "type": "string"
                    }
                },
                "required": [
                    "end",
                    "speaker",
                    "start",
                    "text"
                ]
            },
            "ErrorResponse": {
                "type": "object",
                "properties": {
                    "detail": {
                        "type": "string"
                    }
                },
                "required": [
                    "detail"
                ]
            },
...
            "TranscribeResponse": {
                "type": "object",
                "properties": {
                    "text": {
                        "type": "string"
                    }
                },
                "required": [
                    "text"
                ]
            }
        },
        "securitySchemes": {
            "oauth2": {
                "type": "oauth2",
                "flows": {
                    "clientCredentials": {
                        "tokenUrl": "/o/token/",
                        "scopes": {
                            "read": "Read access to API resources",
                            "write": "Write access (upload, process)"
                        }
                    },
                    "authorizationCode": {
                        "authorizationUrl": "/o/authorize/",
                        "tokenUrl": "/o/token/",
                        "scopes": {
                            "read": "Read access to API resources",
                            "write": "Write access (upload, process)"
                        }
                    }
                }
            }
        }
    }
}